Route for input1 and Conform on HomeInformation
